+++ src/isPrime.h
@@ -1,4 +1,5 @@
 #ifndef _GREATESTPRIME_ISPRIME_H_
 #define _GREATESTPRIME_ISPRIME_H_
-bool isPrime(unsigned long long Number);
+
+bool isPrime(int_fast64_t Number);
 #endif //_GREATESTPRIME_ISPRIME_H_
